Revamp persistent download state handling
Instead of saving the state to a separate `.samfusdl_state` file, the
state will now be saved to a 516-byte block at the end of the download
file. The block contains a parity field and two fixed-size arrays for
storing the remaining download ranges. samfusdl will alternate between
the two arrays for each write of the current state to (hopefully)
prevent unclean computer shutdowns from corrupting the state. All writes
are flushed before and after writing the state.

The state is written every 5 seconds. This is currently a hardcoded
value and cannot be changed via command-line options.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>